backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File handling
aiofiles==23.2.1
python-multipart==0.0.6

# Original dependencies
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
assemblyai==0.23.0

# Enhanced audio processing dependencies
ffmpeg-python==0.2.0
pydub==0.25.1

# Enhanced parallel processing
retrying==1.3.4
psutil==5.9.5

# Enhanced logging
structlog==23.1.0

# Security and file handling
werkzeug==3.0.1

# OpenAI for Medical Extraction (BioBERT removed)
openai==1.12.0

# Core dependencies - Windows compatible versions
numpy==1.26.4
pandas==2.1.4

httpx==0.24.1